fix project_equivalence_properties_test
